Add support for Promises in module (#11)
* Add support for Promises in module

Uses `util.promisify` built into Node v8 to create a Promise from the
is-mongodb-running function.

These changes are backwards compatible with the current callback style
function. If the Promise-style function is desired, simple
require('is-mongodb-running/promise') to get the promisified function
instead.

Updated the is-mongodb-running binary to use promises. Code looks a bit
cleaner

* Fix nits